Draw Constellation plot with scrattch.hicat package.

Draw Constellation plot with scrattch.hicat package.

Usage

constellationPlot(
  SeuratObj,
  cluster_label = "cell_type",
  out.dir = "Constellation_plots",
  k = 15,
  knn.outlier.th = 2,
  outlier.frac.th = 0.5
)

Arguments

SeuratObj

Seurat object

cluster_label

one column of SeuratObj@meta.data, used for cluster annotation to display in the plot.

out.dir

location to write plotting files to

k

K for KNN algorithm.

knn.outlier.th

Threshold to determine if a nearest neighbor is too far

outlier.frac.th

Threshold to determine if a cell is a outlier if it is too far away from most of its neighbors.

Examples

if (FALSE) {
SeuratObj <- pbmc3k.SeuratData::pbmc3k.final
Seurat::Idents(SeuratObj) <- SeuratObj$seurat_clusters
constellationPlot(SeuratObj, cluster_label = "seurat_annotations", out.dir = "Constellation_plots")
}